+++ mayan/apps/authentication/icons.py
@@ -2,5 +2,6 @@ from __future__ import absolute_import, unicode_literals
 
 from mayan.apps.appearance.classes import Icon
 
+icon_impersonate_start = Icon(driver_name='fontawesome', symbol='user')
 icon_logout = Icon(driver_name='fontawesome', symbol='sign-out-alt')
 icon_password_change = Icon(driver_name='fontawesome', symbol='key')