+++ mayan/apps/authentication/literals.py
@@ -2,3 +2,7 @@ from __future__ import unicode_literals
 
 DEFAULT_LOGIN_METHOD = 'username'
 DEFAULT_MAXIMUM_SESSION_LENGTH = 60 * 60 * 24 * 30  # 30 days
+
+IMPERSONATE_VARIABLE_ID = '_impersonate_id'
+IMPERSONATE_VARIABLE_DISABLE = '_impersonate_end'
+IMPERSONATE_VARIABLE_PERMANENT = '_impersonate_permanent'