+++ mayan/apps/authentication/permissions.py
@@ -1,0 +1,13 @@
+from __future__ import absolute_import, unicode_literals
+
+from django.utils.translation import ugettext_lazy as _
+
+from mayan.apps.permissions import PermissionNamespace
+
+namespace = PermissionNamespace(
+    label=_('Authentication'), name='authentication'
+)
+
+permission_users_impersonate = namespace.add_permission(
+    label=_('Impersonate users'), name='users_impersonate'
+)