+++ HISTORY.rst
@@ -11,6 +11,8 @@
 - Increase the size of the file cache maximum size field.
 - Remove codecov dependency.
 - Remove pathlib2 dependency, it is now part of the standard Python library.
+- Remove Django's admindocs app
+- Add user impersonation support.
 
 3.3.15 (2020-03-05)
 ===================